create PanelDisplay component
The intention was for each <PanelDisplay> to handle its corresponding
legend and styling. One (slight) regression here is that the legends
can now scroll off the screen but this can be fixed at a later date
and was only a "nice-to-have".
